+++ fsl_mrs/scripts/fsl_mrs_preproc.py
@@ -66,6 +66,10 @@ def main():
         metavar=('<lower-limit>', '<upper-limit>'),
         default=(0.2, 4.2),
         help='ppm limits of alignment window (default=0.2->4.2)')
+    optional.add_argument(
+        '--align_window',
+        type=int,
+        help='Alignment window size. Disabled by default, but useful for low SNR data.')
     optional.add_argument('--hlsvd', action="store_true",
                           help='Apply HLSVD for residual water removal.')
     optional.add_argument('--leftshift', type=int, metavar='POINTS',
@@ -263,7 +267,12 @@ def main():
                 no_prewhiten=no_prewhiten)
 
     verbose_print('... Align Dynamics (1st iteration) ...')
-    supp_data = nifti_mrs_proc.align(supp_data, 'DIM_DYN', ppmlim=args.align_limits, report=report_dir)
+    supp_data = nifti_mrs_proc.align(
+        supp_data,
+        'DIM_DYN',
+        ppmlim=args.align_limits,
+        window=args.align_window,
+        report=report_dir)
 
     # Bad average removal on the suppressed data
     if args.unlike:
@@ -291,7 +300,12 @@ def main():
 
     # Frequency and phase align the FIDs
     verbose_print('... Align Dynamics (2nd iteration) ...')
-    supp_data = nifti_mrs_proc.align(supp_data, 'DIM_DYN', ppmlim=args.align_limits, report=report_dir)
+    supp_data = nifti_mrs_proc.align(
+        supp_data,
+        'DIM_DYN',
+        ppmlim=args.align_limits,
+        window=args.align_window,
+        report=report_dir)
 
     if 'DIM_DYN' in ref_data.dim_tags:
         ref_data = nifti_mrs_proc.align(ref_data, 'DIM_DYN', ppmlim=(0, 8))